refactor code to fix clippy warnings
