+++ src/parse/typescript.rs
@@ -1,4 +1,4 @@
-use crate::ast::ast::Expression;
+use crate::ast::base::Expression;
 use crate::ast::import::{ImportDeclaration, ImportSpecifier};
 use nom::branch::alt;
 // use nom::branch::alt;
